Add SwiftUI example UI tests
